==> seurat_workflow/5b_annotation_label_transfer.Rmd <==
---
title: "Label Transfer - `r commandArgs(trailingOnly = TRUE)[4]`"
description:
author: "Aaron Allen, original from Devika Agarwal"
date: 'Last update: `r date()`'
output:
  html_document:
    number_sections: true
    code_folding: show
    code_download: true
    theme: cerulean
    df_print: paged
    fig_width: 8.5
    fig_height: 5
    toc: true
    toc_float:
      collapsed: false
      smooth_scroll: false
    toc_depth: 6
---




```{css, echo = FALSE}
pre, code {white-space:pre !important; overflow-x:auto}
```


```{r echo = FALSE, cache = FALSE}

```


```{r setup, echo = FALSE, cache = FALSE}
ggplot2::theme_set(ggplot2::theme_grey())

rmd_name <- gsub(pattern = ".Rmd", replacement = "", x = knitr::current_input(), ignore.case = TRUE)
knitr::opts_chunk$set(dev = c("png", "cairo_pdf"),
                      fig.align = "center",
                      fig.height = 5,
                      fig.width = 8.5,
                      dpi = 300,
                      pdf.options(encoding = "ISOLatin9.enc"),
                      fig.path = paste0("../analyses/figures/", rmd_name, "-",
                                        format(Sys.time(), "%Y%m%d_%H%M%S"), "/"),
                      fig.retina = 1,
                      warning = TRUE,
                      message = TRUE)
```




# Setup

## Job_ID = `r commandArgs(trailingOnly=T)[2]`


## Start time

```{r start-time}
set.seed(12021)
today <- format(Sys.time(), "%Y%m%d_%H%M%S")
total_start_time <- Sys.time()
total_start_time
```







## Parameters



```{r params}
params_file <- commandArgs(trailingOnly = TRUE)[3]
params_path <- paste0("../src/param_files/", params_file)
print(paste0("The params file is :   ", params_path))
```

```{r, class.source = 'fold-hide', code = readLines(params_path)}
```





## Libraries

```{r packages-etc, warning = FALSE, message = FALSE}
library(Seurat)
library(tidyverse)
library(cowplot)
library(pheatmap)
library(colorspace)
library(Matrix)
library(future)
```



```{r run-time-tibble}
process_time <- tibble()
stop_the_clock <- function(chunk_name = "unnamed") {
    temp <- tibble::tibble(Code_chunk = chunk_name,
                           Start_time = start_time,
                           Stop_time = Sys.time(),
                           Run_time = difftime(Stop_time, Start_time, units = "mins")
                           )
    process_time <- dplyr::bind_rows(process_time, temp)
    return(process_time)
}
```



# The data

## Load data


First load our data object of interest.



```{r load-data}
start_time <- Sys.time()
if (remove_doublets == FALSE && remove_ambient == FALSE) {
    path_modifier <- ""
}
if (remove_doublets == TRUE && remove_ambient == FALSE) {
    path_modifier <- "_doublets_removed"
}
if (remove_doublets == FALSE && remove_ambient == TRUE) {
    path_modifier <- "_ambient_removed"
}
if (remove_doublets == TRUE && remove_ambient == TRUE) {
    path_modifier <- "_doublets_removed_ambient_removed"
}

integrated <- read_rds(file = paste0(objects_path,
                                    dataset, "_",
                                    input_type, "_",
                                    normalization, "_",
                                    integration, "_",
                                    "integrated_seurat",
                                    path_modifier, "_",
                                    "annotated.rds")
                            )

process_time <- stop_the_clock("load_data")
```



Get the name of the machine that this is running on.  If running on my local desktop ("mentok"), then append to the `reference_object_path` variable.

```{r is-local}
local_path_mod <- ""
curr_machine <- Sys.info()["nodename"]
if (curr_machine == "mentok") {
    local_path_mod <- "/mnt/data2/scRNAseq/cbrg"
}
local_path_mod
```

```{r adjust-path}
reference_object_path <- paste0(local_path_mod, reference_object_path)
reference_object_path
```


Second, load the reference dataset.

```{r load-ref}
start_time <- Sys.time()
reference_object <- commandArgs(trailingOnly = TRUE)[4]
reference_name <- reference_object %>% str_remove(pattern = ".rds")
reference_object <- read_rds(file = paste0(reference_object_path, reference_object))
DefaultAssay(reference_object) <- "RNA"
reference_name
reference_object
process_time <- stop_the_clock("load_ref")
```


# Quick DimPlot


```{r, fig.width= 20,fig.height=10}
pt_size <- AutoPointSize(data = integrated, raster = FALSE)
DimPlot(object = integrated,
        reduction = paste0("umap_", curr_pc, "pcs"),
        group.by = "merged_clusters",
        label = TRUE,
        label.box = TRUE,
        label.size = 3,
        pt.size = pt_size,
        raster = FALSE) +
    NoAxes() +
    coord_fixed()
```



```{r, fig.width= 20,fig.height=10}
pt_size <- AutoPointSize(data = reference_object, raster = FALSE)
DimPlot(object = reference_object,
        reduction = "tsne",
        group.by = "cell_type",
        label = TRUE,
        label.box = TRUE,
        label.size = 3,
        pt.size = pt_size,
        raster = FALSE) +
    NoAxes() +
    coord_fixed()
```







# Transfer data v1


## PCA project RNA assay


```{r find-anchors}
start_time <- Sys.time()
allgenes <- rownames(integrated@assays$RNA)
var_features_ref <- reference_object@assays$RNA@var.features
intersection_var <- intersect(allgenes, var_features_ref)
DefaultAssay(integrated) <- "RNA"
anchors_ref <- FindTransferAnchors(reference = reference_object,
                                 query = integrated,
                                 dims = 1:curr_pc,
                                 reference.reduction = "pca",
                                 reference.assay = "RNA",
                                 query.assay = "RNA",
                                 normalization.method = "LogNormalize",
                                 features = intersection_var,
                                 max.features = 200,
                                 k.filter = 200)
process_time <- stop_the_clock("find_anchors")
```



### Anchors QC

Info from Devika:

* % of query cells with anchors
* The Seurat reference-mapping procedure first identifies a set of 'anchors', or pairwise correspondences between cells predicted to be in a similar biological state, between query and reference datasets.
* Here we report the percentage of query cells participating in an anchor correspondence.
* Typically, we observe values >15% when mapping is successful.
* Processing mismatched biological datasets (i.e. mapping a query dataset of human brain cells onto a reference dataset of human blood cells) will return few anchors (<5%).
* However, in some cases, when there is a large batch effect between query and reference datasets from the same tissue, this metric can fall below 15% even when mapping is successful

```{r}
anchors <- as.data.frame(anchors_ref@anchors)
query_unique <- length(unique(anchors$cell2))
percent_anchors <- round(query_unique / ncol(integrated) * 100, digits = 2)
cat("% of query dataset cells with anchors:", paste0(percent_anchors, "%"), fill = TRUE)
```



### Mapping Scores

Info from Devika:

* Mapping scores: This value from 0 to 1 reflects confidence that this cell is well represented by the reference.

* The mapping score is independent of a specific annotation, is calculated using the MappingScore function in Seurat, and reflects how well the unique structure of a cell's local neighborhood is preserved during reference mapping.

* A cell can get a low prediction score because its probability is equally split between two clusters (which lowers the prediction score), but the mapping score will remain high.

* A high prediction score means that a high proportion of reference cells near a query cell have the same label. However, these reference cells may not represent the query cell well, resulting in a low mapping score. Cell types that are not present in the reference should have lower mapping scores.


```{r map-score}
start_time <- Sys.time()
mapping_scores_rna_both <- MappingScore(anchors = anchors_ref, ndim = curr_pc)
integrated <- AddMetaData(object = integrated,
                        metadata = mapping_scores_rna_both,
                        col.name = paste0("label_transfer_mapscores_", reference_name))
process_time <- stop_the_clock("map_score")
```


### Transfer labels

 + plot original celltype, predicted  board celltype,  Max prediction score for broad celltypes, Mapping scores based on the transfer anchors

```{r transfer-ann}
start_time <- Sys.time()
predictions_celltype <- TransferData(anchorset = anchors_ref,
                                     refdata = reference_object$cell_type,
                                     dims = 1:curr_pc)
integrated <- AddMetaData(object = integrated,
                          metadata = predictions_celltype$predicted.id,
                          col.name = paste0("label_transfer_cell_type_", reference_name))
integrated <- AddMetaData(object = integrated,
                          metadata = predictions_celltype$prediction.score.max,
                          col.name = paste0("label_transfer_pred_score_max_cell_type_", reference_name))
process_time <- stop_the_clock("transfer_ann")
```


```{r, fig.width=10, fig.height=10}
pt_size <- AutoPointSize(data = integrated, raster = FALSE)
p1 <- DimPlot(object = integrated,
        reduction = paste0("umap_", curr_pc, "pcs"),
        group.by = "merged_clusters",
        pt.size = pt_size,
        repel = TRUE,
        label = TRUE,
        label.box = TRUE) +
    # NoLegend() +
    coord_fixed()
my_legend <- get_legend(p1)
p1 <- p1 + theme(legend.position = "none", validate = TRUE)
plot(p1)
plot(my_legend)
p1 <- DimPlot(object = integrated,
        reduction = paste0("umap_", curr_pc, "pcs"),
        group.by = paste0("label_transfer_cell_type_", reference_name),
        pt.size = pt_size,
        label = TRUE,
        repel = TRUE,
        label.box = TRUE) +
    # NoLegend() +
    coord_fixed()
my_legend <- get_legend(p1)
p1 <- p1 + theme(legend.position = "none", validate = TRUE)
plot(p1)
plot(my_legend)
FeaturePlot(object = integrated,
        reduction = paste0("umap_", curr_pc, "pcs"),
        features = paste0("label_transfer_pred_score_max_cell_type_", reference_name),
        pt.size = pt_size,
        order = TRUE,
        cols = c("lightsteelblue2", "black")) +
    # NoLegend() +
    coord_fixed()
FeaturePlot(object = integrated,
        reduction = paste0("umap_", curr_pc, "pcs"),
        features = paste0("label_transfer_mapscores_", reference_name),
        pt.size = pt_size,
        order = TRUE,
        cols = c("lightsteelblue2", "black")) +
    # NoLegend() +
    coord_fixed()
```


### Transfer labels : Annotations part 2

+ Transfer Annotation labels by using the query *pca* rather than *pcaproject* for the weight reduction


```{r transfer-ann-2}
start_time <- Sys.time()
DefaultAssay(integrated) <- "RNA"
predictions_wrpca_annotation <- TransferData(anchorset = anchors_ref,
                                             refdata = reference_object$cell_type,
                                             dims = 1:curr_pc,
                                             weight.reduction = integrated[["harmony"]])

integrated <- AddMetaData(object = integrated,
                          metadata = predictions_wrpca_annotation$predicted.id,
                          col.name = paste0("label_transfer_cell_type_wrpca_", reference_name))
integrated <- AddMetaData(object = integrated,
                          metadata = predictions_wrpca_annotation$prediction.score.max,
                          col.name = paste0("label_transfer_pred_score_max_cell_type_wrpca_", reference_name))
process_time <- stop_the_clock("transfer_ann_2")
```

+ plot original celltype, predicted  borad cell type,  Max prediction score for broad celltypes, Mapping scores for whole brain data based on the transfer anchors



```{r, fig.width=10, fig.height=10}
pt_size <- AutoPointSize(data = integrated, raster = FALSE)
p1 <- DimPlot(object = integrated,
        reduction = paste0("umap_", curr_pc, "pcs"),
        group.by = "merged_clusters",
        pt.size = pt_size,
        repel = TRUE,
        label = TRUE,
        label.box = TRUE) +
    # NoLegend() +
    coord_fixed()
my_legend <- get_legend(p1)
p1 <- p1 + theme(legend.position = "none", validate = TRUE)
plot(p1)
plot(my_legend)
p1 <- DimPlot(object = integrated,
        reduction = paste0("umap_", curr_pc, "pcs"),
        group.by = paste0("label_transfer_cell_type_wrpca_", reference_name),
        pt.size = pt_size,
        label = TRUE,
        repel = TRUE,
        label.box = TRUE) +
    coord_fixed()
my_legend <- get_legend(p1)
p1 <- p1 + theme(legend.position = "none", validate = TRUE)
plot(p1)
plot(my_legend)
FeaturePlot(object = integrated,
        reduction = paste0("umap_", curr_pc, "pcs"),
        features = paste0("label_transfer_pred_score_max_cell_type_wrpca_", reference_name),
        pt.size = pt_size,
        order = TRUE,
        cols = c("lightsteelblue2", "black")) +
    # NoLegend() +
    coord_fixed()
FeaturePlot(object = integrated,
        reduction = paste0("umap_", curr_pc, "pcs"),
        features = paste0("label_transfer_mapscores_", reference_name),
        pt.size = pt_size,
        order = TRUE,
        cols = c("lightsteelblue2", "black")) +
    # NoLegend() +
    coord_fixed()
```






# Save data


```{r save-end}
start_time <- Sys.time()
write_rds(x = integrated,
          file = paste0(objects_path,
                      dataset, "_",
                      input_type, "_",
                      normalization, "_",
                      integration, "_",
                      "integrated_seurat",
                      path_modifier, "_",
                      "annotated.rds")
          )

dir.create(paste0(objects_path, "/transfer_labels/"), recursive = TRUE, showWarnings = FALSE)
write_rds(x = mapping_scores_rna_both,
          file = paste0(objects_path, "/transfer_labels/label_transfer_mapscores_", reference_name, ".rds"))
write_rds(x = predictions_celltype,
          file = paste0(objects_path, "/transfer_labels/label_transfer_cell_type_", reference_name, ".rds"))
write_rds(x = predictions_wrpca_annotation,
          file = paste0(objects_path, "/transfer_labels/label_transfer_cell_type_wrpca_", reference_name, ".rds"))

process_time <- stop_the_clock("save_end")
```







# Run time

```{r run-times}
process_time
```

```{r plot-run-times, fig.width=10, fig.height=5}
process_time$Code_chunk <- factor(process_time$Code_chunk, levels = rev(process_time$Code_chunk))
process_time %>%
    ggplot(aes(x = Code_chunk, y = Run_time, fill = Run_time)) +
    geom_col(width = 0.8) +
    xlab("Code Chunk") +
    ylab("Run Time (mins)") +
    coord_flip()
```



```{r stop-time}
total_end_time <- Sys.time()
total_end_time
total_end_time - total_start_time
```



# Session info

```{r session-info}
sessionInfo()
```
