==> seurat_workflow/6a_manual_split_and_save.Rmd <==
---
title: "6a_manual_split_and_save"
description: description
author: "Aaron M. Allen"
date: 'Last update: `r date()`'
output:
  html_document:
    number_sections: true
    code_folding: show
    code_download: true
    theme: cerulean
    df_print: paged
    fig_width: 8.5
    fig_height: 5
    toc: true
    toc_float:
      collapsed: false
      smooth_scroll: false
    toc_depth: 6
---



```{css, echo = FALSE}
pre, code {white-space:pre !important; overflow-x:auto}
.tocify-item {white-space:pre}
```



```{r echo = FALSE, cache = FALSE}

```


```{r setup, echo = FALSE, cache = FALSE}
ggplot2::theme_set(ggplot2::theme_grey())
ggplot2::theme_update(panel.background = ggplot2::element_rect(fill = "transparent", colour = NA),
                      plot.background = ggplot2::element_rect(fill = "transparent", colour = NA))

rmd_name <- gsub(pattern = ".Rmd", replacement = "", x = knitr::current_input(), ignore.case = TRUE)
knitr::opts_chunk$set(
    # dev = c("cairo_pdf"),
    dev = c("png", "cairo_pdf"),
    dev.args=list(bg="transparent"),
    fig.align = "center",
    fig.height = 5,
    fig.width = 8.5,
    dpi = 100,
    pdf.options(encoding = "ISOLatin9.enc"),
    fig.path = paste0("../analyses/figures/", rmd_name, "-",
                    format(Sys.time(), "%Y%m%d_%H%M%S"), "/"),
    fig.retina = 1,
    warning = TRUE,
    message = TRUE)
```



# Setup


## Start time

```{r start-time}
set.seed(12021)
today <- format(Sys.time(), "%Y%m%d_%H%M%S")
total_start_time <- Sys.time()
total_start_time
```



## Libraries

```{r packages-etc, warning = FALSE, message = FALSE}
library(Seurat)
library(tidyverse)
library(colorspace)
library(cowplot)
```










# Load data



```{r load-meta-brain}
meta_cb <- read_rds("../../proj136/analyses/rds_files/New_Meta_cbNeuron_PUBnofacs_noPolIII_v1_alevin_LogNormalize_harmony_integrated_seurat_doublets_removed_diet.rds")
meta_cb
```





## Subcluster


```{r}
subcluster_by <- "annotation_broad_cell_type_ext_fix"
```

```{r}
meta_cb@meta.data[[subcluster_by]] %>% unique() %>% sort()
```



```{r}


meta_cb@meta.data[[subcluster_by]] <- meta_cb@meta.data[[subcluster_by]] %>% as.character() %>% replace_na("unknown")
split_meta_cb <- SplitObject(object = meta_cb, split.by = subcluster_by)
for (i in seq_along(split_meta_cb)) {

	message(paste0(names(split_meta_cb)[[i]], ":"))
    
	message(paste0("\t ... splitting"))
	subcluster_list <- SplitObject(object = split_meta_cb[[i]], split.by = "orig.ident")
	
    message(paste0("\t ... cleaning"))
	subcluster_list_clean <- list()
	for (ii in seq_along(subcluster_list)) {
	    counts_to_use <- "RNA"
	    subcluster_list_clean[[ii]] <- CreateSeuratObject(counts = subcluster_list[[ii]]@assays[[counts_to_use]], project = names(subcluster_list)[[ii]])
		subcluster_list_clean[[ii]] <- AddMetaData(object = subcluster_list_clean[[ii]], metadata = subcluster_list[[ii]]$orig.ident, col.name = "orig.ident")
		subcluster_list_clean[[ii]] <- AddMetaData(object = subcluster_list_clean[[ii]], metadata = subcluster_list[[ii]]$experiment, col.name = "experiment")
		subcluster_list_clean[[ii]] <- AddMetaData(object = subcluster_list_clean[[ii]], metadata = subcluster_list[[ii]]$tissue, col.name = "tissue")
		subcluster_list_clean[[ii]] <- AddMetaData(object = subcluster_list_clean[[ii]], metadata = subcluster_list[[ii]]$cell_nuclei, col.name = "cell_nuclei")
		subcluster_list_clean[[ii]] <- AddMetaData(object = subcluster_list_clean[[ii]], metadata = subcluster_list[[ii]]$availability, col.name = "availability")
		subcluster_list_clean[[ii]] <- AddMetaData(object = subcluster_list_clean[[ii]], metadata = subcluster_list[[ii]]$sex, col.name = "sex")
		subcluster_list_clean[[ii]] <- AddMetaData(object = subcluster_list_clean[[ii]], metadata = subcluster_list[[ii]]$percent_mt, col.name = "percent_mt")
		subcluster_list_clean[[ii]] <- AddMetaData(object = subcluster_list_clean[[ii]], metadata = subcluster_list[[ii]]$percent_rRNA, col.name = "percent_rRNA")
		subcluster_list_clean[[ii]] <- AddMetaData(object = subcluster_list_clean[[ii]], metadata = subcluster_list[[ii]]$percent_rProt, col.name = "percent_rProt")
		subcluster_list_clean[[ii]] <- AddMetaData(object = subcluster_list_clean[[ii]], metadata = subcluster_list[[ii]]$percent_hsp, col.name = "percent_hsp")
	}
	names(subcluster_list_clean) <- names(subcluster_list)
	subcluster_list_clean


    message(paste0("\t ... saving"))
    write_rds(x = subcluster_list, 
        file = paste0(
				"../../proj136/analyses/rds_files/", 
				"New_Meta_cbNeuron_PUBnofacs_noPolIII_v1--1000VarGene--NoReg--Harmony_ori_exp_cell--",  
				"_alevin_seurat_list_doublets_removed_", 
				names(split_meta_cb)[[i]], "_", 
				"subclustering.rds")
        )
}

```





```{r}
list.files(path = "../../proj136/analyses/rds_files/", pattern = "New_Meta_cbNeuron_PUBnofacs_noPolIII_v1--1000VarGene--NoReg--Harmony_ori_exp_cell--")
```













# Session info

```{r session_info}
sessionInfo()
```
